refactor: move map into separate component
